Rename the configuration file to a sample version
Configuration of Sensu may take place exclusively using the `conf.d` directory.
The addition of the original `config.json` file will occur upon upgrading the
Sensu omnibus package after a systems administrator had intentionally deleted
it. The restart of Sensu services will merge the stock configuration found in
the `config.json` with other configuration. This is not desired and the manual
removal of `config.json` is inconvenient after upgrade. This change ensures
that the package does not place a configuration file that would be loaded by
default. This change requires a documentation update to explain the possible
adoption of `config.json.sample` for new installations.